Fix some warnings in dod.c. Patch courtesy of Kevin Tew <tewk@tewk.com> (who has also been added to CREDITS).
git-svn-id: https://svn.parrot.org/parrot/trunk@9659 d31e2699-5ff4-0310-a27c-f18f2fbe73fe

+++ src/dod.c
@@ -296,7 +296,7 @@ Parrot_dod_trace_root(Interp *interpreter, int trace_stack)
     /* Now, go run through the PMC registers and mark them as live */
     /* First mark the current set. */
     ctx = CONTEXT(interpreter->ctx);
-    for (i = 0; i < ctx->n_regs_used[REGNO_PMC]; i++) {
+    for (i = 0; i < (unsigned int)ctx->n_regs_used[REGNO_PMC]; i++) {
         if (REG_PMC(i)) {
             pobject_lives(interpreter, (PObj *)REG_PMC(i));
         }
@@ -509,7 +509,7 @@ trace_active_buffers(Interp *interpreter)
     /* First mark the current set. We assume that all pointers in S registers
      * are pointing to valid buffers. This is not a good assumption, but it'll
      * do for now. */
-    for (i = 0; i < CONTEXT(interpreter->ctx)->n_regs_used[REGNO_STR]; i++) {
+    for (i = 0; i < (unsigned int)CONTEXT(interpreter->ctx)->n_regs_used[REGNO_STR]; i++) {
         Buffer *reg = (Buffer *)REG_STR(i);
 
         if (reg)
